requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
pydantic
pydantic-settings
python-dotenv
